# HBASE-24831 Avoid invoke Counter using reflection in SnapshotInputFormat