Avoid invoke Counter using reflection in SnapshotInputFormat