@utf7 2020-08-07
with the pr we get about 7.6%/9.1% performance improve 


in my test case with ScanPerformanceEvaluation tools

like that 

`hbase  org.apache.hadoop.hbase.ScanPerformanceEvaluation -t snapshotscanmapreduce  -rs /tmp/hbase/rs1 -sn sp_test`

**with pr**

1#

total time to scan: 143943 ms
total bytes: 16357785600 bytes (15.2 G)
throughput  : 109.1 MB/s
total rows  : 10485760
throughput  : 71.6 K rows/s
total cells : 104857600
throughput  : 716.1 K cells/s


2#

total time to scan: 145948 ms
total bytes: 16357785600 bytes (15.2 G)
throughput  : 107.6 MB/s
total rows  : 10485760
throughput  : 70.6 K rows/s
total cells : 104857600
throughput  : 706.2 K cells/s

total time to scan: 149962 ms
total bytes: 16357785600 bytes (15.2 G)
throughput  : 104.7 MB/s
total rows  : 10485760
throughput  : 68.7 K rows/s
total cells : 104857600
throughput  : 687.2 K cells/s



**without pr**

1th:
total time to scan: 156929 ms
total bytes: 16357785600 bytes (15.2 G)
throughput  : 100 MB/s
total rows  : 10485760
throughput  : 65.6 K rows/s
total cells : 104857600
throughput  : 656.4 K cells/s

2th
total time to scan: 156078 ms
total bytes: 16357785600 bytes (15.2 G)
throughput  : 100 MB/s
total rows  : 10485760
throughput  : 65.6 K rows/s
total cells : 104857600
throughput  : 656.4 K cells/s